changes file to tsx
